fix: close IDOR in createNotificationForUser; route cross-user notifies through internal helper
CodeRabbit r3870528952 (CWE-639): createNotificationForUser authenticated
the caller but never authorized the target username, so any authenticated
client could inject notifications into another user's feed by calling the
exported server action directly.

- Split notifications into ./internal (no "use server"): notifyUser writes a
  notification and is NOT a portable server action / not client-callable.
  Trusted server flows that already established the caller-to-target
  relationship (share, reminders, mentions, assignment) now import
  notifyUser from there.
- createNotificationForUser exported action now authorizes the target:
  self or admin only. Cross-user notifications from non-admin callers are
  blocked at the action boundary.
- KanbanCardDetail: remove the direct client-side createNotificationForUser
  call for assignment (the actual IDOR vector). The assignment notification
  now happens server-side in checklist-item updateItem, which already
  verifies currentUser + canReach(EDIT) before notifying via notifyUser.
- Tests: update mock paths to ./internal; add IDOR and self-notify tests.
