+++ app/_server/actions/notifications/index.ts
@@ -1,68 +1,12 @@
 "use server";
 
-import path from "path";
 import { AppNotification, AppNotificationData } from "@/app/_types";
 import { Modes, NotificationTargets } from "@/app/_types/enums";
-import { NOTIFICATIONS_FILE } from "@/app/_consts/files";
 import { getCurrentUser } from "@/app/_server/actions/users";
-import { readJsonFile, writeJsonFile, ensureDir } from "@/app/_server/actions/file";
-import { broadcast } from "@/app/_server/actions/ws/broadcast";
 import { getListById } from "@/app/_server/actions/checklist";
 import { getNoteById } from "@/app/_server/actions/note";
 import { mountsFor } from "@/app/_server/actions/share/mounts";
-
-const DEDUP_WINDOW_MS = 60 * 60 * 1000;
-
-const _getPath = (username: string): string => NOTIFICATIONS_FILE(username);
-
-const _read = async (username: string): Promise<AppNotification[]> => {
-  const data = await readJsonFile(_getPath(username));
-  return Array.isArray(data) ? data : [];
-};
-
-const _write = async (username: string, notifications: AppNotification[]): Promise<void> => {
-  const filePath = _getPath(username);
-  await ensureDir(path.join(process.cwd(), path.dirname(filePath)));
-  await writeJsonFile(notifications, filePath);
-};
-
-const _isDuplicate = (
-  existing: AppNotification[],
-  type: AppNotification["type"],
-  data?: AppNotificationData,
-): boolean => {
-  if (type === "assignment") return false;
-
-  const cutoff = Date.now() - DEDUP_WINDOW_MS;
-  const key =
-    type === "reminder"
-      ? data?.taskId
-      : type === "mention"
-        ? data?.commentId
-        : data?.itemId;
-  if (!key) return false;
-
-  return existing.some((n) => {
-    if (n.type !== type) return false;
-    if (new Date(n.createdAt).getTime() <= cutoff) return false;
-    const nKey =
-      n.type === "reminder"
-        ? n.data?.taskId
-        : n.type === "mention"
-          ? n.data?.commentId
-          : n.data?.itemId;
-    return nKey === key;
-  });
-};
-
-const _buildNotification = (
-  data: Omit<AppNotification, "id" | "createdAt">,
-): AppNotification => ({
-  ...data,
-  link: undefined,
-  id: `${Date.now()}-${Math.random().toString(36).slice(2, 9)}`,
-  createdAt: new Date().toISOString(),
-});
+import { notifyUser, readNotificationsForUser, writeNotificationsForUser } from "./internal";
 
 const CATEGORY_MODES: Partial<Record<NotificationTargets, Modes>> = {
   [NotificationTargets.NOTE_CATEGORY]: Modes.NOTES,
@@ -110,24 +54,19 @@ const _resolveLink = async (
   return undefined;
 };
 
+/** Authorizes the target: self or admin only. Cross-user flows use `notifyUser`. */
 export const createNotificationForUser = async (
   username: string,
   data: Omit<AppNotification, "id" | "createdAt" | "link">,
 ): Promise<{ success: boolean }> => {
-  try {
-    const caller = await getCurrentUser();
-    if (!caller?.username) return { success: false };
+  const caller = await getCurrentUser();
+  if (!caller?.username) return { success: false };
 
-    const existing = await _read(username);
-    if (_isDuplicate(existing, data.type, data.data)) return { success: true };
-
-    await _write(username, [_buildNotification(data), ...existing]);
-    await broadcast({ type: "notification", action: "created", username });
-    return { success: true };
-  } catch (error) {
-    console.error("[notifications] createNotificationForUser failed:", error);
+  if (caller.username !== username && !caller.isAdmin) {
     return { success: false };
   }
+
+  return notifyUser(username, data);
 };
 
 export const createNotification = async (
@@ -135,14 +74,14 @@ export const createNotification = async (
 ): Promise<{ success: boolean }> => {
   const user = await getCurrentUser();
   if (!user?.username) return { success: false };
-  return createNotificationForUser(user.username, data);
+  return notifyUser(user.username, data);
 };
 
 export const getNotifications = async (): Promise<AppNotification[]> => {
   const user = await getCurrentUser();
   if (!user?.username) return [];
 
-  const notifications = await _read(user.username);
+  const notifications = await readNotificationsForUser(user.username);
 
   return Promise.all(
     notifications.map(async (n) => ({
@@ -156,8 +95,8 @@ export const markNotificationRead = async (id: string): Promise<{ success: boole
   try {
     const user = await getCurrentUser();
     if (!user?.username) return { success: false };
-    const notifications = await _read(user.username);
-    await _write(user.username, notifications.map((n) =>
+    const notifications = await readNotificationsForUser(user.username);
+    await writeNotificationsForUser(user.username, notifications.map((n) =>
       n.id === id ? { ...n, readAt: new Date().toISOString() } : n,
     ));
     return { success: true };
@@ -171,8 +110,8 @@ export const markAllNotificationsRead = async (): Promise<{ success: boolean }> 
     const user = await getCurrentUser();
     if (!user?.username) return { success: false };
     const readAt = new Date().toISOString();
-    const notifications = await _read(user.username);
-    await _write(user.username, notifications.map((n) => (n.readAt ? n : { ...n, readAt })));
+    const notifications = await readNotificationsForUser(user.username);
+    await writeNotificationsForUser(user.username, notifications.map((n) => (n.readAt ? n : { ...n, readAt })));
     return { success: true };
   } catch {
     return { success: false };
@@ -183,8 +122,8 @@ export const removeNotification = async (id: string): Promise<{ success: boolean
   try {
     const user = await getCurrentUser();
     if (!user?.username) return { success: false };
-    const notifications = await _read(user.username);
-    await _write(user.username, notifications.filter((n) => n.id !== id));
+    const notifications = await readNotificationsForUser(user.username);
+    await writeNotificationsForUser(user.username, notifications.filter((n) => n.id !== id));
     return { success: true };
   } catch {
     return { success: false };
@@ -195,7 +134,7 @@ export const clearNotifications = async (): Promise<{ success: boolean }> => {
   try {
     const user = await getCurrentUser();
     if (!user?.username) return { success: false };
-    await _write(user.username, []);
+    await writeNotificationsForUser(user.username, []);
     return { success: true };
   } catch {
     return { success: false };
